fix comments and separated toggle button to different component
