+++ client/src/components/App/AppToolbar/TopToolbar/IsActiveToggle/IsActiveToggleStyles.ts
@@ -1,0 +1,16 @@
+import {makeStyles} from "@material-ui/styles";
+
+const useStyles = makeStyles({
+    toggle: {
+        borderRadius: '2.5vh',
+        height: '3vh',
+        width: '4.6vw',
+        fontSize: '0.9vw',
+    },
+    isActiveToggle: {
+        cursor: 'pointer',
+        marginRight: '3vw',
+    }
+});
+
+export default useStyles;
