+++ client/src/components/App/AppToolbar/TopToolbar/TopToolbarStyles.ts
@@ -23,16 +23,6 @@ const useStyles = makeStyles({
     centering: {
         alignSelf: 'center'
     },
-    toggle: {
-        borderRadius: '2.5vh',
-        height: '3vh',
-        width: '4.6vw',
-        fontSize: '0.9vw',
-    },
-    isActiveToggle: {
-        cursor: 'pointer',
-        marginRight: '1vw',
-    },
     swalTitle: {
         fontSize: '1.5vw',
         fontFamily: 'Assistant',
